Make DC work better with the macro recorder.  Specifically, 4 changes.
1. Remove the custom DC results table.  Instead use the main results
table (Analyzer.getResultsTable()) and move the logic to add/remove
detected circles into the togglecircle tool.

2. Change the class name from 'Run_' to 'Detect_Circles'. This way the
auto generated macro code looks like this

run("Detect Circles", ...);

rather than this

run("Run ", ...);

3. Replace the spaces with underscores in the GenericDialog field names.
This is a requirement for the macro recorder.

4. Don't automatically activate the circle toggle tool after the detect
circle code runs.  It seems that this code was confusing the macro
recorder.  With that code, the macro recorder was only capturing the
command to activate the circle toggle tool and not the command to run
the plugin.